implement method `Clear()`, and improve locking (#11)
This change adds the method Clear() to bloom filter.

Additonally, it fixes an earlier flaw in the locking, when a union was made, the filter needs to not only lock the filter being written to, but also the filter being read from. When doing so, we also need to be careful about the case when both arguments are the same: otherwise we wind up in a deadlock.